rust: syn: import crate
This is a subset of the Rust `syn` crate,
version 1.0.104, licensed under "Apache-2.0 OR MIT", from:

    https://github.com/dtolnay/syn/raw/1.0.104/src

The files are copied as-is, with no modifications whatsoever
(not even adding the SPDX identifiers).

For copyright details, please see:

    https://github.com/dtolnay/syn/blob/1.0.104/README.md#license
    https://github.com/dtolnay/syn/blob/1.0.104/LICENSE-APACHE
    https://github.com/dtolnay/syn/blob/1.0.104/LICENSE-MIT

The next two patches modify these files as needed for use within
the kernel. This patch split allows reviewers to double-check
the import and to clearly see the differences introduced.

The following script may be used to verify the contents:

    for path in $(cd rust/syn/ && find . -type f -name '*.rs'); do
        curl --silent --show-error --location \
            https://github.com/dtolnay/syn/raw/1.0.104/src/$path \
            | diff --unified rust/syn/$path - && echo $path: OK
    done

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>