rust: serde_cbor: import crate
This is a subset of the Rust `serde_cbor` crate,
version v0.11.2, licensed under "Apache-2.0 OR MIT", from:

   https://github.com/pyfisch/cbor/tree/v0.11.2/src

The files are copied as-is, with no modifications whatsoever
(not even adding the SPDX identifiers).

For copyright details, please see:

   https://github.com/pyfisch/cbor/tree/v0.11.2#license
   https://github.com/pyfisch/cbor/blob/v0.11.2/LICENSE-APACHE
   https://github.com/pyfisch/cbor/blob/v0.11.2/LICENSE-MIT

The next patch modifies these files as needed for use within
the kernel. This patch split allows reviewers to double-check
the import and to clearly see the differences introduced.

The following script may be used to verify the contents:

   for path in $(cd rust/serde_cbor/ && find . -type f -name '*.rs'); do
       curl --silent --show-error --location \
           https://github.com/pyfisch/cbor/raw/v0.11.2/src/$path \
           | diff --unified rust/serde_cbor/$path - && echo $path: OK
   done