rust: hex: implement FromHex trait and hex::decode using a custom kernel_alloc feature
This commit also adds a process_results function for Vec<Result<T,E>>
which transforms it into a Result<Vec<T,E>>, collecting all the values
inside the Result elements into a new vector which directly stores the
elements.
For this to work, the TryReserveError was added to the FromHexError
enum.
This is probably not the ideal design, but since collect cannot be used
directly, it's a good substitute.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>